[#28] Fix Attachments MVVM pattern
Replace [AttachmentVM] with [Attachment], because it doesn't match
MVVM design pattern.
