--- MKVStudio/ViewModels/Main/BaseViewModel.cs
@@ -1,10 +1,0 @@
-﻿using PropertyChanged;
-using System.ComponentModel;
-
-namespace MKVStudio.ViewModels;
-
-[AddINotifyPropertyChangedInterface]
-public class BaseViewModel : INotifyPropertyChanged
-{
-    public event PropertyChangedEventHandler PropertyChanged = (sender, e) => { };
-}
